requirements.txt
bidict==0.22.1
blinker==1.6.2
click==8.1.3
Flask==2.3.2
Flask-SocketIO==5.3.4
h11==0.14.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
python-engineio==4.4.1
python-socketio==5.8.0
RPi.GPIO==0.7.1
simple-websocket==0.10.0
Werkzeug==2.3.4
wsproto==1.2.0
